Bug 50579 - Error count is long, sample count is int. Changed sample count to long.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@1059242 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: ef0f94a58e3b4f46556dba2b7d10b3dfd81d937a

+++ src/components/org/apache/jmeter/visualizers/DistributionGraph.java
@@ -168,7 +168,7 @@ public class DistributionGraph extends JComponent implements Scrollable, Clearab
             Number ninety = p_model.getPercentPoint(0.90);
             Number fifty = p_model.getPercentPoint(0.50);
 
-            int total = p_model.getCount();
+            long total = p_model.getCount();
             Collection<Number[]> values = p_model.getDistribution().values();
             Number[][] objval = values.toArray(new Number[][]{});
             // we sort the objects
